Let trill spanners end before accidental of next note
This is logical (the trill ends on the note and the accidental is part
of the note), and recommended by Gould (Behind Bars p. 137).

To implement this, extend the line-spanner-interface's logic to find
accidentals so it also works if the bound is not a note head or a rest
(i.e. something contained in a note column) but the note column
itself.  This obviates DurationLine's own logic for that; this has to
be removed in order not to get double adjustments.

This also includes a fix to the Trill_spanner_engraver: there is no
reason not to set the bound of the old trill spanner to the note
column if there is a new one.  Having a paper column instead of a note
column doesn't give the accidentals, which would defeat this change on
consecutive trill spanners.