Allow printing the first bar number without \bar ""
From the user point of view, this gets rid of a command that is
necessary just for technical reasons.

From the technical perspective, the larger goal behind this is to
reimplement \bar with a stream event since that will allow to rewrite
the interaction between bar lines and line breaks in a better way.
This means that the underlying implementation will likely be reworked,
but doing this allows more gradual changes.  This implementation just
initializes the boolean that tells whether a bar line has been found
to #t at the beginning of the piece.

Regression tests show interesting differences: it appears that a
number of people believed all-bar-numbers-visible was among the
relevant settings to make bar numbers printed in the middle of
systems, whereas just BarNumber.break-visibility suffices.  This makes
the behavior just as expected, but to help resolve the confusion, a
warning is added to Documentation/en/changes.tely and in convert-ly.