Add \enablePolymeter command
\enablePolymeter replaces

\context {
  \Score
  \remove Timing_translator
  \remove Default_bar_line_engraver
}
\context {
  \Staff
  \consists Timing_translator
  \consists Default_bar_line_engraver
}

It also works with Staff-like contexts, such as TabStaff.

One goal is to make the input more convenient.  Another is to be able
to rename the Default_bar_line_engraver and/or add more engravers to
the set that has to be moved for polymetric notation with less
backwards compatibility issues.