Add Staff alias to the context definition of Dynamics
The Dynamics context is somewhat special because it is inserted within
containers that typically hold staves (Score, StaffGroup, etc.), but
is also a bottom context.  Giving it both Staff and Voice aliases
seems like a good way to reflect this peculiarity.  This fixes
\magnifyStaff not working in Dynamics (because there was no context
with name or alias Staff above a Dynamics context), and will ease the
implementation of a better user interface for polymetric music.

Closes #6188