Implement new pre-process-music translator slot
pre-process-music is called on all translators after listeners but
before process-music.  It allows processing that needs all events
heard but should write context properties early so that other
translators can read these properties in process-music without
dependency on engraver order.  Similarly to the existing two-phase
listeners+process-music and acknowledgers+process-acknowledged steps,
it provides a kind of 'semantic' hook: it allows engravers written by
different people to more often than not work together seamlessly if
they conform to the principle that pre-process-music is used for
processing considered preliminary, and process-music for the bulk of
the action of creating grobs.  Future commits will use it in
Paper_column_engraver (obviating \overrideProperty for paper columns).