Suggest convert-ly when compilation fails with new version
When the compilation results in an error, and the version of LilyPond
used to compile is more recent than the \version statement, print a
warning to encourage use of convert-ly.  This should hopefully raise
awareness about the existence of this tool, reducing the number of
questions about why newer versions fail to compile files with old
syntax, and generally improve the user experience with respect to
backwards compatibility by teaching people who don't realize that the
tool exists in the first place.

There is no warning when compilation succeeds because it would be
rather noisy, and it's legitimate to try to write code compatible with
several versions.

A "more recent" version is defined by having a higher major version
number; or a higher minor number; or equal major and minor numbers and
higher patch number, but only if it's a development release (odd minor
number) because syntax does not change in stable releases, so the
warning would be useless noise in that case.

Closes #3784