Add events for volta-style repeat start and end
This is good because it encapsulates the setting of the repeatCommands
context property inside Repeat_acknowledge_engraver (aside from its use
for manual repeats, which is already documented with a caveat).

This also increases the similarity between segno-style and volta-style
handling.

The start event tells the repeat count.  The end event tells a repeat
count and the return count.  These are not used in this commit.