Let `lilymidi --pretty` output velocity values
For checking the action of dynamic controls, it is important to be
able to see the velocity values with which notes are struck.  At the
current point of time, LilyPond does not give non-zero velocity values
to note-off events but `lilymidi` is also useful for looking at MIDI
files not created by LilyPond so any non-zero velocity is reported.

The way velocities are reported is by appending `@15` for a velocity
of 15 (for example).