Web: update Windows support statement
It was severely outdated, promising support for Windows versions as
old as Windows 2000, for which we (as well as Microsoft) provide zero
support these days.  The phrase "current versions of Windows" doesn't
promise too much regarding support while not preventing people from
trying and maybe finding that it works in a certain configuration.
See
https://lists.gnu.org/archive/html/lilypond-devel/2022-02/msg00088.html