Ensure ASCII names for files passed to Ghostscript
As both the input PS file as well as the output PDF / PNG file(s) are
temporary, there is no benefit in using the final basename in their
name and we can as well use only ASCII characters. This avoids problems
with Ghostscript on Windows and fixes the second part of issue #6282.
While at it, completely move the temporary PS file to /tmp since its
contents are copied, not moved, to the final destination.