side-position-interface: revise handling of nonexistent skylines
Remove the warning.  Empty skylines can be useful, if only for
understanding what skylines are for and debugging.  There is no harm
to this: the default skylines are extent-based, not empty, so there is
no risk that a developer would give an object empty skylines by
accident.

Fix a crash.  The default, an empty skyline, always had its direction
set to UP by the argumentless Skyline constructor, which led to a
crash with an assertion failure if the direction was UP.