Trade _ for G_ in all Scheme code
Translatable messages are now marked with G_ instead of _, following
the guidance from the Guile documentation
(https://www.gnu.org/software/guile/manual/html_node/Gettext-Support.html).
The problem with _ is that it conflicts with the syntactic keyword
used in syntax-case, match and such. In Guile 2 and later, macros
defined with the syntax-case macro system recognize syntactic
keywords, but only if they are unbound (cf. the footnote at
https://www.gnu.org/software/guile/manual/html_node/Syntax-Rules.html#DOCF14).
This means that, for example, trying to use _ as the wildcard in
(ice-9 match) pattern matching will not work as expected.

The interesting changes are in scm/lily.scm and stepmake/stepmake/po-targets.make.
The rest was made mechanically using:

  git ls-files "*.scm" "*.ly" "*.tely" "*.itely" "*.texi" "*.itexi" | xargs sed -i "s/(_ /(G_ /g"
  git restore Documentation/en/contributor/programming-work.itexi

(programming-work.itexi reverted because it's talking about C++ code;
I have verified all other occurrences for being correct.  One instance
of (_"...") without space between _ and " has been fixed in
define-markup-commands.scm.)