Deduplicate Bézier code through Scheme interfaces to C++
There was duplication of functionality to compute the extent of a
Bézier curve between the Bezier C++ class and stencil.scm.  The Scheme
version was causing trouble in Guile 2 for certain edge cases in
paths.  As the C++ version is the one used for ties and slurs, it is
more thoroughly tested with Guile 2, so use that one, adding
specializations of is_scm, from_scm and to_scm to allow converting
Bézier curves between C++ and Scheme.

Closes #6291