Refactor SRFI-17 setters
Guile 3 has trouble with using define (or its variant define-public)
to set the value of a variable that already exists.  There is
apparently some kind of shadowing where uses of the variable happening
before the redefinition try to refer to the later definition when
compiled, which obviously fails.  Instead, define the setters in C++
directly using a new macro, LY_DEFINE_WITH_SETTER.