IR: reduce table indentation in PDF output
In the IR, the length of (almost) all first columns in a two-column table is
too long to allow

```
foo:   bar
```

formatting.  Instead, we have

```
foobarbaz:
       bar
```

For this reason, reduce the table indentation of the second column in PDF
output to be identical to the indentation in `@itemize` environments.  As a
result, we get a longer line length, which reduces the number of 'underfull
line' warnings and shortens the length of the IR by approx. 3% (currently,
this is a reduction from 753 to 734 pages).