break-substitution.cc: use SCM_UNSPECIFIED instead of SCM_UNDEFINED
Sometimes, there is no clear way to define what should happen when
break substitution is applied to a data structure and some grob it
contains does not have a sibling on the system in question.  Use
SCM_UNSPECIFIED in these cases because SCM_UNDEFINED is an internal C
API value without a meaning in Scheme.

Closes #6222