break-substitution.cc: use a simple loop instead of goto
The change in behavior is that *unspecified* is no longer removed from
the list if a grob has no sibling on the system.  This is saner
because in nested pairs there is no way in general to remove the
orphan, and it should not matter anyway since the change is only for
lists containing grobs, which should normally not be used as we have
Grob_array for that.