Remove filter_map_assign optimization in break substitution
The intent was to reuse the memory of the grob array returned by the
first grob substitution, with item directions, for the second
substitution, with actual systems.  The optimization does not work
very well because the new array is not necessarily the same size as
the old.  Below are benchmarks for this commit.

les-nereides.ly
===============

CPU type: Intel(R) Core(TM) i5-8265U CPU @ 1.60GHz
CPU frequency set to 1950 MHz
52 runs

                     Time summary
+----------------------------------------------------+
|                | Baseline | New executable | Delta |
|----------------+----------+----------------+-------|
| Min            | 1.430    | 1.450          | 1.40% |
| Mean           | 1.477    | 1.488          | 0.79% |
| Standard error | 0.003    | 0.002          |       |
+----------------------------------------------------+

Difference of means is 4.72 times standard error.

                    Memory summary
+-----------------------------------------------------+
|                | Baseline  | New executable | Delta |
|----------------+-----------+----------------+-------|
| Min            | 92456.000 | 93556.000      | 1.19% |
| Mean           | 92777.077 | 93846.538      | 1.15% |
| Standard error | 24.491    | 21.699         |       |
+-----------------------------------------------------+

Difference of means is 46.31 times standard error.

mozart-hrn-3.ly
===============

20 runs

                     Time summary
+-----------------------------------------------------+
|                | Baseline | New executable | Delta  |
|----------------+----------+----------------+--------|
| Min            | 4.140    | 4.000          | -3.38% |
| Mean           | 4.217    | 4.078          | -3.30% |
| Standard error | 0.010    | 0.011          |        |
+-----------------------------------------------------+

Difference of means is 13.49 times standard error.

                     Memory summary
+------------------------------------------------------+
|                | Baseline   | New executable | Delta |
|----------------+------------+----------------+-------|
| Min            | 144396.000 | 145148.000     | 0.52% |
| Mean           | 144771.000 | 145444.400     | 0.47% |
| Standard error | 53.152     | 42.732         |       |
+------------------------------------------------------+

Difference of means is 14.05 times standard error.

MSDM.ly
=======

21 runs

                     Time summary
+-----------------------------------------------------+
|                | Baseline | New executable | Delta  |
|----------------+----------+----------------+--------|
| Min            | 36.140   | 35.300         | -2.32% |
| Mean           | 36.417   | 35.766         | -1.79% |
| Standard error | 0.042    | 0.055          |        |
+-----------------------------------------------------+

Difference of means is 13.51 times standard error.

                     Memory summary
+-------------------------------------------------------+
|                | Baseline    | New executable | Delta |
|----------------+-------------+----------------+-------|
| Min            | 1187216.000 | 1203316.000    | 1.36% |
| Mean           | 1188121.333 | 1204171.238    | 1.35% |
| Standard error | 100.960     | 110.623        |       |
+-------------------------------------------------------+

Difference of means is 151.71 times standard error.

------

The general conclusion is that removing the optimization results in
some speed increase, which is marked for mozart-hrn-3.ly (3.5%) and to
a lesser extent for MSDM.ly (2%), even though the smaller test
les-nereides.ly has a decrease (1%).  On the memory side, this commit
is fatter by about 1%.  Arguing that compilation time is an important
factor in the typesetting of any imposing project whereas memory is
usually not an issue except for extremely large scores, this change
should have positive net payoff from the performance point of view,
even if not accounting for reduced code complexity.