Accept incomplete version statements within stable releases
Commit 0ecab848b9 inadvertently changed behavior by rejecting
incomplete version statements like \version "1" or \version "2.22".
On the other hand, convert-ly has always rejected such versions.  This
brings consistency by accepting partial versions iff they have two
digits, and the second is even.  This is because it can make sense to
have \version "2.22", as syntax is supposed not to change within a
stable release, whereas \version "2.23" is not useful to convert-ly
because the syntax does change within an unstable release.