cairo; set bounding box explicitly for EPS output
Before, in EPS mode, Cairo dumped a tight bounding box around the
drawing extents, eating the indent of the first line of music in
lilypond-book output.

Set the BoundingBox and PageBoundingBox explicitly. This makes the EPS
output of the regtest interchangeable with the traditional PS backend,
save for (sub)pixel rendering details.