Improve skyline Scheme functions
Having these as callbacks on the Skyline class was confusing, because
normally that mechanism is used for grob callbacks.  These are not
callbacks of a 'skyline interface'.  This converts them to LY_DEFINE
and a more usual naming Scheme.  While at it, also add a few missing
type checks, write docstrings, rename arguments so they make more
sense in the IR, use from_scm/to_scm.