Remove anchored links from news headlines
The anchor for the unstable version had been broken since the release
of LilyPond 2.21.4 when it started saying "LilyPond-2_002e21_004e3"
instead of "LilyPond-2_002e21_002e4" (where _002e encodes a dot). The
anchor for the stable version had been broken since the release of
LilyPond 2.20 when it started missing a _0021 to encode the exclamation
mark.
I corrected this in commit 058e184edf while looking at the situation for
the release of LilyPond 2.23.7, but given this is a major pain to update
and has been broken for so long without anybody complaining, let's just
remove it.

Closes #6006