Prevent fixscm.sh from badly reformatting a few forms
A match form should be formatted as

(match subject
  (pattern1 expr1)
  (pattern2 expr2)
  ...)

and not

(match subject
       (pattern1 expr1)
       (pattern2 expr2)
       ...)

Same goes for a few other forms.

This is implemented via .dir-locals.el, so those developers who use
Emacs also get the right behavior out-of-the-box.

This has also been submitted as a patch to Emacs, which has been
merged and will thus be included in the next release:
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=55033