Use double colon for some Scheme comments
Comments taking a whole line, as opposed to comments in the margin of
a code line, should use double colons.  Emacs rigidly enforces the
convention that line comments are started with double colons, and
margin comments with single colons.  As a result, fixscm.sh indents
single-colon comments very far to the right.