Grob::internal_get_object: remove a check for unpure/pure containers
This check made little sense: there is no such thing as purity for
grob objects.  The effect of this commit is that if LilyPond encounters
an unpure/pure container as grob object property, it will no longer
ignore it (giving '()) but just return it.  Since unpure/pure
containers are never put in the object_alist_, this makes no
difference.