Allow replacing the Score context
With these changes, it is possible to define a context type with \type
Score_engraver, make Global accept it, and instantiate it via \new, or
make it the \defaultchild of Global, just like this is possible for
contexts that do not have the special status of Score as the
descendant of Global.

Among motivating use cases, you can define a ProportionalScore that is
a Score with proportional notation turned on.  This is the example
used in the regtests and the added documentation.

Implementation-wise, this moves the code that applies output def
settings from Output_def::instantiate to the Context constructor.
This constructor is called by the derived Global_context constructor.
Consequently, Global_context settings are no longer hardcoded.