Extend staff to 'staff-bar aligned grob
Previously, staff lines would extend to a grob with bar-line-interface,
but that was too specific.  Other grobs can be break-aligned in the
'staff-bar position to serve a similar purpose without being a true
bar line; it is reasonable to print the staff as if they were bar lines.

Without this change, the staff lines would stop too early in the new
regression test.

There is also a change to Note_spacing made for consistency at the code
level.  I am not sure how to test whether it really matters.