Obviate second argument of LY_DEFINE (Scheme function name)
LY_DEFINE was checking that the Scheme function name corresponded to
the C++ name.  This goes one step further and removes the second
argument, letting it be computed automatically from the first.  This
is exactly what was already done with MAKE_SCHEME_CALLBACK.

The interesting parts to review are in
lily/include/lily-guile-macros.hh and lily/function-documentation.cc.
The rest is an automated removal of this argument in all instances of
the macro.