Trill_spanner_engraver: inline typeset_all ()
It was needless to call it in finalize (), since finalize () comes
right after a stop_translation_timestep () and typeset_all () always
left finished_ being null.

Relatedly, add a regression test for trills being implicitly
terminated by the end of the score, without \stopTrillSpan.  The
output of this test is not affected by the rest of this commit.