Make TrillSpanner use true extents
Previously, trill lines could exceed the allotted line length due to
the trill_element sticking out of its bounding box (it is so because
two trill elements should blend together when stacked horizontally
with zero padding).  This makes them always remain within the length,
and gives them their actual extent, not the mere extent of the
combined trill elements.  The "Tr" symbol gets true extents, too.
Changes to extents require slightly adapting the TrillSpanner
definition to adjust the spacing between "Tr" and the first trill
element (this is actually a simplification).

The purpose of this commit is to allow introducing collision detection
for consecutive trill spanners, which requires accurate extents.