Improve formatting of chained trill spanners
This avoids the situation where, due to the integer number of
squiggles, two consecutive trills can be printed on the same vertical
position or stacked depending on the precise horizontal spacing
(stacked trill spanners when shortening the first one by one squiggle
would have allowed chaining them look ugly).  To this end, the
Trill_spanner_engraver recognizes this situation and sets the
right-neighbor object of the first trill to the second trill.  The
line-spanner-interface gains an adjust-on-neighbor subproperty of
bound-details.{left,right} that reads this object to adapt the length.

Fixes #923