Let \break force a break in all circumstances, add \allowBreak
The Paper_column_engraver no longer ignores \break if something else
prevented a break.  Rather, it understands that the user really wants
a break, and bypasses other considerations.  It sets a new property,
forceBreak, when it receives such an event requesting a break
explicitly.  This is done in pre-process-music so that it can be used
by other translators in process-music.  The new function
break_allowed() checks "forceBreak or (not forbidBreak)".  This is
used as a replacement for acknowledging the bar-line-interface in a
number of engravers, to allow clefs, key signatures, etc. to be
printed at the beginning of a line even if there was no bar line at
that point.

A side benefit is making \remove Bar_engraver more useful, since it
will no longer suppress clefs etc. on the second line and following.

Engravers can no longer conclude that a break won't be allowed upon
not seeing a BarLine, nor upon seeing that forbidBreak is #t.  They
should use break_allowed() instead.

Further more, taking advantage of the fact that Paper_column_engraver
no longer disregards explicit breaking commands when forbidBreak is
set, this adds \allowBreak, which inserts a break point whatever
conditions might have led LilyPond not to allow one by default.  [This
is not a separate commit because it's needed for some of the
regtests.]

Fixes #6243