Doc-LM: remove a warning about \new Score
It seems that there were misunderstandings: we prefer writing \score
instead of \new Score in snippets because \score is more powerful
(allows adding \layout and \midi), but there is no technical reason
why \new Score would be bad.  \new SomeScoreLikeContext is going to be
the recommended syntax for instantiating a score-level context when
this feature is added in a later commit, so there would be some
dissonance.