Let PercentRepeat use centered-spanner-interface
The logic of PercentRepeat was repeating that of the
centered-spanner-interface, so now that this interface is able to
handle grobs whose stencil isn't just their text property interpreted,
it is refactored to share it.  This means that the
percent-repeat-item-interface is no longer relevant, since
PercentRepeat is a spanner, so this eliminates it, moving everything
to the percent-repeat-interface.