Revert "Obviate second argument of LY_DEFINE (Scheme function name)"
Having the Scheme function name in the source code makes it possible
to easily grep for the identifier and it also simplifies the generation
of tags files for IDE support.

This reverts most of commit a67f14a0cfca2b908c0585c95f1f93977f459c74,
but retains the documentation in lily-guile-macros.hh and adds the
Scheme function name for ly_skylines_for_stencil.