Let \alterBroken really support callbacks
The regtest had a test case for callbacks in the list of property
values passed to \alterBroken, but the support was actually
accidental.  What happened is that on the first time the property was
looked up, the callback generated by \alterBroken returned the
callback given by the user, which would appear as the value of the
property.  On a second lookup, that user callback would be resolved.
This led to inconsistencies (for example, in the regtest,
point-and-click didn't work on the hairpin, because its extents used
the wrongly computed value).  This lets value-for-spanner-piece
execute the value as a callback if it's a procedure.  Support for
unpure/pure containers is also added, which is especially useful with
the grob-transformer.