Remove -dgui option
This option had two effects: It redirected all output to a .log file
and, if lilypond was started without input files, opened an editor
with the file Welcome_to_LilyPond.ly.
As far as I can tell, the only real use of this flag was on Windows
where the installer produced by GUB created shortcuts using this
option. For users wishing to redirect the output message, it is far
more convenient to use their shell's functionality for this.