Add line break to bar-line-manual-v-repeat test
Clobbering ".|:" with "" at a line break is worth covering because it is
perhaps unintuitive that "" should be able to clobber a visible bar, and
more so in this case, since the one isn't present at the end of the line
and the other isn't present at the beginning.

I have lately tried changing this behavior to address issue 3688, but
the subtleties of my preferred alternative are dissatisfying and
converting existing scores would be tricky, so I expect the status quo
(\bar overrides automatic bar lines) to continue.