Stop using global header to make a score title when normal title is empty
Previously, there was a bit of code that, upon seeing that the score
title was empty, tried to get a score title using only the variables
from the global \header block, without the ones from the \score.
There is no comment explaining the necessity, no regtest affected by
removing it, and no discernible rationale for it.  It might be an
artifact of refactorings in output def variable inheritance.  At any
rate, it leads to confusing behavior where turning off more title
fields in \score can lead to more being printed, so just remove it.

Fixes #6359