Improve get_translator_group error path
Use ly_scm_write_string instead of ly_symbol2scm to make this work if
the provided type is not a symbol.  This is the case for '(), which is
the value if no \type was given.  This is a somewhat easy mistake to
make, and it would previously error out with

ERROR: In procedure apply-smob/2:
In procedure symbol->string: Wrong type argument in position 1 (expecting symbol): ()

Also remove dead code (error () does not return).