translation-functions.scm: Update figured bass formatter
By using Unicode values we automatically get the new text-mode accidentals;
it is also no longer necessary to enforce a uniform glyph height due to the
recent change of aligning the stack elements of the figured bass on
baselines.

Also reduce the size of the figured bass to a more standard (i.e., smaller)
value.