binaries: Fix linking of libintl on macOS
Because the scripts build static libraries, the CoreFoundation
framework needs to be explicitly specified in the LDFLAGS. This
used to work with GLib 2.70, but was broken in 2.71.1. It never
worked for building LilyPond, but I failed to notice that this
optional dependency was missing, which seems actually required
to use filenames with special characters.

Fixes #6362