Define "x" bar lines for future use with \caesura
It has long been possible to define a bar line with a mid-line stencil
but no beginning-of-line or end-of-line stencils, but the reverse has
not been possible.

This adds bar types "x-|", "x-||", and "x-." to support styles that
allow line breaks between lyric phrases (whether or not they fall on
measure boundaries) and use a specific bar line for such a break.

I've been less thorough in testing this than I usually am.  A bunch of
expected uses will be covered with the coming support for \caesura in
Bar_engraver, but there are other obvious cases worthy of attention:

* (define-bar-line "x" #f #f #f)
* "x" or "x-..." in the EOL/BOL/span arguments of bar-line definitions
* bar types with other glyph components combined with "x"
* tests of layering that are independent of \caesura

I don't want to record here which of the above I currently think should
be allowed and which should be disallowed, but it's some of both.