Layer bar lines per break direction
Instead of choosing a single glyph string and using it for beginning-,
middle-, and end-of-line grobs, look up a glyph string for each case
independently.

This simplifies combining automatic bar lines: for many common cases, it
is no longer necessary to use annotated bar lines.

To support this, it was necessary to redefine these bar types so that
they no longer demand a bar line at the end of a line:

    .|    .|:    [|:    S    S.|:

This commit provides some internal logic anticipating a \noBar command,
but it does not provide the command.