Make \caesura generally available
\caesura is available by default, not just when including gregorian.ly.

\caesura also no longer overrides BreathingSign properties, which allows
them to be set differently in different contexts.

breathing-sign-ancient.ly is revised to adjust to the above changes and
to reduce repetition.  Changes in horizontal spacing are the result of
not overriding BreathingSign properties.

These changes are not advertised in the documentation because in modern
contexts, \caesura produces the same mark as \breathe, but this is
expected to change once Caesura_engraver is implemented.