CG: Drop "Summary of project status" node
The Frog label no longer exists.  It's unclear why the
"Maintainability" label deserves a link but not any other label, and
all important labels are already explained in "Issue classification".
The list of all issues is easy to find in the GitLab interface.  Lists
of MRs with specific ~"Patch::*" labels do not belong to "Issues", and
they aren't very useful as there aren't too many patches at the same
time (these links probably used to be useful during the times of
SourceForge, when each patch was reviewed in an issue, and patches
would have been drown under other issues without labels).