feta-numbers.mf: Fatten thin lines of 'figbass' glyphs considerably
This improves legibility at small sizes; it also makes the figured bass
glyphs look more similar to Urtext editions of Bärenreiter or Henle.

Note that this commit also affects all other number glyphs: The now larger
value of the `thin` parameter necessitates changes of most shapes to achieve
more 'metaness'.  This is done by avoiding or reducing dependencies of point
coordinates on this parameter except points that are directly related to the
thickness of outlines.  However, the differences for the non-fattened shapes
are very small and only visible at larger magnifications.