Doc-LM + Web: update setup instructions
Note that previously, these instructions were duplicated on download
pages and in the learning manual.  This is not possible right now
because the download pages are for 2.22, for which the setup
instructions are completely different from the 2.23 ones. This moves
everything to the LM.  That way, the instructions are still correct on
the current "Download" page, because they link to the 2.22 LM, but the
2.23 LM, which is linked to with the 2.23 binaries, works for those
who want to use 2.23.

The new instructions use Frescobaldi because it is a de-facto
standard.  This fixes #3716.

Also replace with English text in all translations.  Everything
changes, so nothing of value is lost anyway, and it seems particularly
important not to display wrong information regarding the installation
setup.  This is done by moving the installation tutorial to its own
section and file, because it deserves it and because I didn't want to
update the English text in all files for other languages when
responding to review comments.

Fixes #6310, #3716