Issue deprecation warning for pre-2.18 override syntax
Old override (and revert) syntax was

    \override Stem #'transparent = ##t

or

    \override Voice.Stem #'transparent = ##t

This deprecated syntax is now flagged with a warning.  Property paths
used as music function arguments do not work without that dot, and the
inconsistency is not helpful.

This will mostly affect source code not passed through `convert-ly`
since the conversion patterns catch (most?) uses.