feta-numbers.mf: Complete number sets
This commit starts a series of four commits to systematize the access of
digits using OpenType features.

* 'cv47' (character variant 47)

  If on, glyphs 'four' and 'seven' have shorter strokes.  Default is off.

* 'tnum' (tabular figures)

  If off (which is the default), glyphs 'zero' to 'nine' have no left and
  right side bearings.  If on, the glyphs all have the same advance width by
  making the bearings non-zero.

* 'ss01' (stylistic set 1)

  If on, glyphs 'zero' to 'nine' have a fatter design, making them more
  readable at small sizes.  Default is off.

* 'kern' (kerning)

  If on (which is the default), glyphs 'zero' to 'nine' and the punctuation
  glyphs 'plus', 'comma', 'hyphen', and 'period' have pairwise kerning.

This eventually leads to the following feature combinations:

* time signatures: (none)
* figured bass: +cv47, +tnum, +ss01
* fingering: +cv47, +ss01
* fixed-width: +cv47, +tnum, -kern

In `feta-numbers.mf`, provide the necessary glyphs for 'cv47', 'tnum', and
'ss01' ('kern' gets implemented later in the font generation chain):

* normal glyphs
* fixed-width glyphs
* fattened glyphs
* fixed-width, fattened glyphs

For all four groups, also provide alternative digit forms for glyphs 'four'
and 'seven'.

Note that there is actually no new code; the new (and renamed) glyphs are
simply recombinations of the existing shapes and flags that control them.