Fix MIDI title encoding confusion
The code was using ly:encode-string-for-pdf, which is specific to the
PDF format and has nothing to do with MIDI.  If the string can't be
encoded in ASCII, ly:encode-string-for-pdf encodes it in UTF16, then
decodes it in latin1.  After reencoding this in UTF8, the result is a
mess.

Fixes #6389