Doc: add warning about \bar creating a spurious staff
Since \bar has been reimplemented with an event in ce27e1a1, it has
the effect of creating a spurious staff in cases like

{
  \bar "something"
  <<
    \new Staff ...
    \new Staff ...
  >>
}

This syntax is a bit dubious anyway (the \bar has to be in the musics
for the individual staves as long as it's not in the beginning), and
it is not expected to be used by many, so just add a warning.  Having
\bar as an event has advantages (in the future, it should enable to
\tweak a \bar), and restoring the old behavior while keeping \bar as
an event incurs extra complexity for little payoff.

Fixes #6341