Fix page references to labels straddling at a page break
They were handled inconsistently when the book contained several
bookparts because the whole alist was reversed instead of just the
part for the local bookpart, leading to several reversals for the same
alist part.  Merge the two loops, addressing a TODO and incidentally
leading to nothing special needing to be done for ensuring a
consistent order.

Fixes #6301