A general facility to transform strings interpreted in markups
This refactors text replacements in terms of "string transformers".
These are procedures stored in the 'string-transformers property for
markups.  When a string is interpreted and there are transformers,
they are first applied, and it is the result that is interpreted.
Furthermore, the result of the transformer can be any markup, which
allows string replacements to replace strings with arbitrary markups.

\smallCaps now works by installing a string transformer, which makes
it work on arbitrary markup arguments rather than just string.

Fixes #1482