Fix attachment of chord grid lines to initial bar line
At the initial time step, a chord square has to be created in
process-music, as usually there won't be a bar line to trigger it
(what looks like a bar line at the beginning is actually a
SystemStartBar, see the TODOs in engraver-init.ly).  However, if there
is a repeat bar, we need to make it the bound retrospectively instead
of just ignoring it.

Fixes #6391